requirements.txt
discord.py
google-generativeai
numpy
python-dotenv
duckduckgo-search
